% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_har_forest.R
\name{estimate_har_tree}
\alias{estimate_har_tree}
\title{Estimate a HAR tree}
\usage{
estimate_har_tree(data, formula, split.vars, minsize, mtry = 1/3, data.predict)
}
\arguments{
\item{data}{A data frame containing the data to be used for the estimation.}

\item{formula}{A formula object specifying the model to be fitted (only classical HAR model for now).}

\item{split.vars}{A vector of splitting variables.}

\item{minsize}{The minimum number of observations in each leaf.}

\item{mtry}{The number of variables to consider at each split.}

\item{data.predict}{A data frame containing the data to be used for the prediction.}
}
\value{
A list containing the following components:
\item{tree}{A data frame containing the tree structure}
\item{split_vars}{A vector of splitting variables}
\item{predictions}{A data frame containing the predictions}
\item{formula}{The formula used for the model}
}
\description{
Estimate a HAR tree using the honest estimation method.
}
