% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HARforest-package.R
\docType{package}
\name{HARforest-package}
\alias{HARforest}
\alias{HARforest-package}
\title{HARforest: Estimation of local linear models for volatility forecasting}
\description{
We provide the functionality to estimate the panel heterogeneous autoregressive forest introduced in "A forest full of risk forecasts for managing volatility" by Kleen and Tetereva (2022, \doi{10.2139/ssrn.4161957}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/onnokleen/HARforest/}
  \item Report bugs at \url{https://github.com/onnokleen/HARforest/issues}
}

}
\author{
\strong{Maintainer}: Onno Kleen \email{r@onnokleen.de} (\href{https://orcid.org/0000-0003-4731-4640}{ORCID})

}
\keyword{internal}
