% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rv_panel_data}
\alias{rv_panel_data}
\title{Stock returns and financial conditions.}
\format{
A data frame with 16,004 rows and 7 variables:
\describe{
\item{date}{date}
\item{permno}{a permanent stock identifier}
\item{rv_lead_22}{5-minute realized variances}
\item{rv_lag_1}{a dummy for each year/week combination}
\item{rv_lag_5}{National Financial Conditions Index}
\item{rv_lag_22}{National Financial Conditions Index}
\item{vix_lag}{National Financial Conditions Index}#' }
}
\usage{
rv_panel_data
}
\description{
A dataset containing the S&P 500 stock returns and the NFCI
}
\keyword{datasets}
