% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{splitting_criterion_honest_package}
\alias{splitting_criterion_honest_package}
\title{This is the splitting criterion we minimize (i.e., sum of squares)}
\usage{
splitting_criterion_honest_package(split.var, data.fit, data.honest, formula)
}
\description{
This is the splitting criterion we minimize (i.e., sum of squares)
}
\keyword{internal}
